add route product price list for inline editing of product prices
- same filters as original product list
- allows editing purchasePrice, margin and sellingPrice
- auto saves whenever a change occurs

The initial search returns ES search objects, so the template is made to handle both ES and ED (which also means adjusting some components to handle both).
ember-data records are fetched to be able to save the records with their changes, but also react on changes in the template.
Data is shown the moment ES returns a response. ED records are loaded while the user can already see the table.
Some helpers specified inside a model are moved out of it so these can be used by ES results too.